dom/tests/mochitest/dom-level0/test_separate_post_message_queue.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for using separate event queue for post messages during page load behaviors</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();
var tab;
var callOrder = [];
function onDone() {
  tab.close();
  isDeeply(callOrder, ["Runnable1", "Runnable2", "Runnable3", "PostMessage",
    "Runnable1", "Runnable2", "Runnable3", "PostMessage"], "Runnables should be fired prior to PostMessage");
  SimpleTest.finish();
}

SpecialPowers.pushPrefEnv({"set":[["dom.separate_event_queue_for_post_message.enabled", true]]}, function () {
  tab = window.open('file_separate_post_message_queue.html');
  tab.loadCount = 1;
});

</script>
</pre>
</body>
</html>
